Fix incremental task evaluation semantics
This is a bridge commit that shows my first approach to solving som
issues with incremental task evaluation. A subsequent commit makes some
changes to the internal data structures in Previous that provide a
cleaner and more general solution. I think the history is still useful
so I didn't squash them down.

While writing documentation for the new file management/incremental
task evaluation features, I realized that incremental task evaluation
did not have the correct semantics. The problem was that calls to
`.previous` are not scoped within the current task. By this, I mean that
say there are tasks foo and bar and that the defintion of bar looks like

bar := {
    val current = foo.value
    foo.previous match {
        case Some(v) if v == current => // value hasn't changed
        case _ => process(current)
    }
}

The problem is that foo.previous is stored in
effectively (foo / streams).value.cacheDirectory / "previous". This
means that it is completely decoupled from foo. Now, suppose that the
user runs something like:
> set foo := 1
> bar // processes the value 1
> set foo := 2
> foo
> bar // does not process the new value 2 because foo was called, which updates the previous value

This is not an unrealistic scenario and is, in fact, common if the
incremental task evaluation is changed across multiple processing steps.
For example, in the make-clone scripted test, the linkLib task processes
the outputs of the compileLib task. If compileLib is invoked separately
from linkLib, then when we next call linkLib, it might not do anything
even if there was recompilation of objects because the objects hadn't
changed since the last time we called compileLib.

To fix this, I add support for tracking the previous file input and file
output stamps as a Map[String, Seq[(Path, FileStamp)]] and fill this map
for the task that needs the previous stamps. This map can be stored in
the previous cache for the calling task. The mechanism for filling and
storing this map is tedious and error-prone, so I introduced two macro
extension methods for tasks: changedInputFiles and changedOutputFiles.

The reason the data structure is a map is that a task may depend on the
file stamps for multiple tasks. Because sbt doesn't support triply
nested tasks, there was no other way I could figure out to encode this.
I used Map[String, Seq[(Path, FileStamp)]] rather than
Map[Scope, Seq[(Path, FileStamp)]] because I didn't want to write a
JsonFormat for Scope.

I also had logic to ensure that this map was only written to the
previous cache if it actually succeeded. In EvaluateTask, I remove any
results for inputFileDependencyMap and outputFileDependencyMap if there
is no successful result for their respective task scopes. This ensures
that the maps reflect the file stamps of the dependent tasks when the
task last successfully completed. I think this matches the intuitive
expectation of what changed file stamps should mean.

I updated the scripted test for make-clone to explicitly test the kidn
of scenario around compileLib and linkLib. The test changes would have
caused a failure with 1.3.0-RC3. I also added tests that ensure that a
task can depend on the input and output file stamps of multiple tasks.