Add enhanced previous semantics
This commit adds the ability for tasks to independenty track the
previous results of other tasks. It does not change the existing
semantics of calls to `.previous` on a task key.

The motivation for this change is that if we have tasks foo and bar and
the definition of bar looks like:
bar := {
  val current = foo.value
  foo.previous match {
    case Some(v) if v == current =>
    case None => process(current)
  }
}
then we want to ensure that if foo is run independently between calls to
bar, that the process function is called whenever the value of foo has
changed _relative_ to the last time that bar has successfully run.

The idea is fairly straightforward. The Previous cache is modified to be
keyed on two tasks, one is the task whose value is being stored (foo in
the example above) and the other is the task in which the stored task
value is retrieved (bar in the example above). When the two tasks are
the same, the behavior is the same as before.

Currently the previous value for foo might be stored somewhere like:

base_directory/target/streams/_global/_global/foo/previous

Now, if foo is stored with respect to bar, it might be stored in

base_directory/target/streams/_global/_global/bar/previous-dependencies/_global/_gloal/foo/previous

By storing the files this way, it is easy to remove all of the previous
values for the dependencies of a task.

In addition to changing how the files are stored on disk, we have to store
the references in memory differently. A given task can now have multiple
previous references (if, say, two tasks bar and baz both depend on the
previous value). When we complete the results, we first have to collect
all of the successful tasks. Then for each successful task, we find all
of its references. For each of the references, we only complete the
value if the scope in which the task value is used is successful.

With this new approach to previous, we can dramatically simplify the
implementation of incremental file task evaluation since it is no longer
necessary to manually track the file dependency maps.

In the actual implemenation in Previous.scala, there are many places
where we have to cast to ScopedKey[Task[Any]]. This is due to
limitations of ScopedKey and Task being type invariant. These casts are
all safe because we never try to get the value of anything, we only use
the portion of the apis of these types that are independent of the value
type. Structural typing where ScopedKey[Task[_]] gets inferred to
ScopedKey[Task[x]] forSome x is a big part of why we have problems with
type inference.