Add previousInThis api
It is useful to be able to get the previous value with respect to the
current task. In order to preserve the old semantics of `.previous`,
this commit adds a new api `previousInThis` which can be used when we
want the previous value to be stored with respect to the last successful
run of the task in which the previous value is fetched.

I add a scripted test that validates that the scoping of the previous
tasks makes sense and that it works transitively.

I think it might be a good idea to just change the semantics of previous
given that I don't think they really made any sense when previous was
called outside of the defining task. This commit is in case we don't
want to take the risk of breaking builds that somehow relied on the old
behavior. My sense is that previous is not a widely used feature and is
probably most commonly used in the task in which it is defined, e.g.
foo := Def.taskDyn {
  foo.previous match {
    case Some(v) => Def.task(v)
    case _ => Def.task(1)
  }
}.value