Don't sort watch inputs
There seems to be an issue with the sorting of globs and the ordering
doesn't really matter. I think I had added sorting as a very minor
performance optimization for registering directories with swoval in
order. Until there is evidence that it matters, it makes sense to just
stop sorting.

Should fix https://github.com/sbt/sbt/issues/4970.