Add better error reporting
We have seen failures in scripted to create the output file for streams
and it has also been reported in https://github.com/sbt/sbt/issues/5067.
This commit is intended to shine light on whether the failure is caused
by the parent directory being absent, which suggests a race condition in
which its being deleted by a different thread, or if the directory does
exists but still the file can't be created. This will not actually fix
anything and is intended solely to aid debugging.