Only set close classloader for interactive sessions
The purpose of closing the classloaders is to prevent resource leaks in
long running sessions. If a user is running batch
commands, then it's fine if they want to use shutdown hooks or leak
threads.

This commit makes it so that we set the `sbt.classloader.close`
property when we enter shell if it isn't already set. I change the
default value of `SysProp.closeClassLoaders` to be false when the
variable is unset. This prevents us from closing classloaders for
non-interactive sessions.