Bump scalafmt
The latest version of the scalafmt plugin updates the internal version
of coursier that is used for resolving the scalafmt version. I was
having a lot of problems with intellij being unable to resolve the
scalafmt version (https://github.com/scalameta/scalafmt/issues/1415).
I'm hoping that this upgrade will fix that issue.

It also incorporates https://github.com/scalameta/sbt-scalafmt/pull/46
which is a nice improvement.